breakdown reducers into server agents and entries
