+++ tornjak-frontend/src/components/agent-list.component.js
@@ -110,8 +110,8 @@ class AgentList extends Component {
 
 
 const mapStateToProps = (state) => ({
-  globalServerSelected: state.serverInfo.globalServerSelected,
-  globalagentsList: state.serverInfo.globalagentsList,
+  globalServerSelected: state.server.globalServerSelected,
+  globalagentsList: state.agents.globalagentsList,
 })
 
 export default connect(
