+++ tornjak-frontend/src/components/entry-list.component.js
@@ -115,8 +115,8 @@ class EntryList extends Component {
 
 
 const mapStateToProps = (state) => ({
-  globalServerSelected: state.serverInfo.globalServerSelected,
-  globalentriesList: state.serverInfo.globalentriesList
+  globalServerSelected: state.server.globalServerSelected,
+  globalentriesList: state.entries.globalentriesList
 })
 
 export default connect(
