+++ tornjak-frontend/src/components/tornjak-server-info.component.js
@@ -94,7 +94,7 @@ class TornjakServerInfo extends Component {
 }
 
 const mapStateToProps = (state) => ({
-  globalServerSelected: state.serverInfo.globalServerSelected,
+  globalServerSelected: state.server.globalServerSelected,
 })
 
 export default connect(
