+++ tornjak-frontend/src/reducers/index.js
@@ -1,8 +1,12 @@
-import selectReducer from './selectReducer';
+import serverReducer from './serverReducer';
+import agentsReducer from './agentsReducer';
+import entriesReducer from './entriesReducer';
 import {combineReducers} from 'redux';
 
 const allReducers = combineReducers({
-    serverInfo : selectReducer
+    server : serverReducer,
+    agents : agentsReducer,
+    entries : entriesReducer,
 });
 
 export default allReducers;
